@@ -1,6 +1,7 @@
-$:.unshift('.', __FILE__)
+$LOAD_PATH.unshift('.', __FILE__)
 require "lib/xcpretty/ansi"
 require "support/matchers/colors"
 require "fixtures/constants"
 
 include XCPretty::ANSI
+
